+++ CHANGES.md
@@ -1,6 +1,7 @@
 Changes 1.5.1-prerelease:
  * remove tests of ansi_cprng
  * apply change suggestions derived from Claude code
+ * kcapi-enc: add tag processing in decryption without ciphertext for completeness reason
 
 Changes 1.5.0:
  * move all sha* applications to the libexec directory to allow them to coexist with other packages sha* applications - the caller is expected to make a symlink to them